Fix intergrate api add and list pet appointments
